pyproject.toml
[tool.poetry]
name = "core-abao-test"
version = "0.1.0"
description = ""
authors = ["A-baoYang <rubyyang0615@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.1.9"
langchain-openai = "^0.0.7"
openai = "^1.12.0"
esg-toolkit = {git = "https://github.com/ESGTrans/esg_toolkit.git", rev = "main"}
python-dotenv = "^1.0.1"
llama-index = "^0.10.12"
matplotlib = "^3.8.3"
llama-index-vector-stores-qdrant = "^0.1.3"
ipython = "^8.22.1"
llama-index-embeddings-clip = "^0.1.4"
clip = {git = "https://github.com/openai/CLIP.git"}
tiktoken = "^0.6.0"
langchain-experimental = "^0.0.53"
simsimd = "3.7.7"
spacy = "^3.7.4"
lxml = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
